added DrawerContent component and added to Drawer folder for the slide view List
