+++ Components/Details.js
@@ -3,18 +3,18 @@ import {View, Text, StyleSheet, Button} from 'react-native';
 import {useFocusEffect, useNavigationState} from '@react-navigation/native';
 
 const DetailsScreen = ({navigation}) => {
-  const state = useNavigationState((state) => state);
+  // const state = useNavigationState((state) => state);
 
-  console.log('Na', state);
+  // console.log('Na', state);
 
-  useFocusEffect(
-    React.useCallback(() => {
-      fetch('https://restcountries.eu/rest/v2/callingcode/372')
-        .then((response) => response.json())
-        .then((data) => console.log(data));
-      return () => console.log('lost focus');
-    }),
-  );
+  // useFocusEffect(
+  //   React.useCallback(() => {
+  //     fetch('https://restcountries.eu/rest/v2/callingcode/372')
+  //       .then((response) => response.json())
+  //       .then((data) => console.log(data));
+  //     return () => console.log('lost focus');
+  //   }),
+  // );
   return (
     <View style={styles.container}>
       <Text style={styles.text}>Details</Text>
