--- Drawer/Contact copy.js
@@ -1,23 +1,0 @@
-import React from 'react';
-import {View, Text, StyleSheet} from 'react-native';
-
-const Contact = () => {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.text}>Contact</Text>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  text: {
-    fontSize: 25,
-  },
-});
-
-export default Contact;
