+++ Drawer/Drawer.js
@@ -8,11 +8,13 @@ import Favorites from './Favorites';
 import ProfileDrawer from './ProfileDrawer';
 import Settings from './Settings';
 
+import DrawerContent from './DrawerContent';
+
 const Drawer = createDrawerNavigator();
 
 const MyDrawer = () => {
   return (
-    <Drawer.Navigator>
+    <Drawer.Navigator drawerContent={(props) => <DrawerContent {...props} />}>
       <Drawer.Screen name="Home" component={MyStack} />
       <Drawer.Screen name="Contact" component={Contact} />
       <Drawer.Screen name="Favorites" component={Favorites} />
