+++ Drawer/DrawerContent.js
@@ -1,0 +1,41 @@
+import React from 'react';
+import {View, StyleSheet} from 'react-native';
+
+import {
+  Avatar,
+  Title,
+  Caption,
+  Paragraph,
+  Drawer,
+  Text,
+  TouchableRipple,
+  Switch,
+} from 'react-native-paper';
+
+import {
+  createDrawerNavigator,
+  DrawerContentScrollView,
+  DrawerItemList,
+  DrawerItem,
+} from '@react-navigation/drawer';
+
+const DrawerContent = (props) => {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.text}>Drawer content</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 25,
+  },
+});
+
+export default DrawerContent;
